Factor out BundleGraphExplorerPane
