+++ app/src/app/App.css
@@ -1,25 +1,3 @@
 body {
     margin: 0;
 }
-
-.container {
-    position: relative;
-    min-height: 100vh;
-}
-
-.sidebar {
-    position: absolute;
-    top: 0;
-    left: 0;
-    background-color: rgba(255, 255, 255, 0.5);
-    border: 1px solid #aaaaaa;
-    z-index: 1;
-}
-
-.graph {
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-}
