+++ app/src/app/BundleGraphExplorerPane.css
@@ -1,0 +1,21 @@
+.container {
+    position: relative;
+    min-height: 100vh;
+}
+
+.sidebar {
+    position: absolute;
+    top: 0;
+    left: 0;
+    background-color: rgba(255, 255, 255, 0.5);
+    border: 1px solid #aaaaaa;
+    z-index: 1;
+}
+
+.graph {
+    position: absolute;
+    top: 0;
+    left: 0;
+    bottom: 0;
+    right: 0;
+}
